Move Dense to base components
