+++ scvi/nn/_base_components.py
@@ -5,6 +5,8 @@ import torch
 from torch import nn as nn
 from torch.distributions import Normal
 from torch.nn import ModuleList
+from flax import linen as lnn
+from flax.linen.initializers import variance_scaling
 
 from scvi._compat import Literal
 
@@ -18,6 +20,14 @@ def reparameterize_gaussian(mu, var):
 def identity(x):
     return x
 
+class Dense(lnn.Dense):
+    def __init__(self, *args, **kwargs):
+        # scale set to reimplement pytorch init
+        scale = 1 / 3
+        kernel_init = variance_scaling(scale, "fan_in", "uniform")
+        # bias init can't see input shape so don't include here
+        kwargs.update({"kernel_init": kernel_init})
+        super().__init__(*args, **kwargs)
 
 class FCLayers(nn.Module):
     """
